+++ src/Serializer/Encoder/HtmlEncoder.php
@@ -25,7 +25,7 @@ class HtmlEncoder implements EncoderInterface
 	 */
 	public function encode($data, $format, array $context = [])
 	{
-		if (!is_string($data) && !is_array($data)) {
+		if (false === is_string($data) && false === is_array($data)) {
 			throw new UnexpectedValueException(sprintf('The HtmlEncoder data must be a array or string or object implementing __toString(), "%s" given.', gettype($data)));
 		}
 		
@@ -58,7 +58,7 @@ class HtmlEncoder implements EncoderInterface
 					$block .= $this->encodeBlock($value);
 					$block .= self::HTML_END_TAG_LIST_ITEM;
 				} else {
-					if (!is_string($value) && !is_numeric($value)) {
+					if (false === is_string($value) && false === is_numeric($value) && false === (is_object($value) && method_exists($value, '__toString'))) {
 						throw new UnexpectedValueException(sprintf('The HtmlEncoder $value must be number or string or object implementing __toString(), "%s" given.', gettype($value)));
 					}